OCR Panneaux/Code/App/back_detection/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
ultralytics
requests
